+++ app/Http/Controllers/Fleet/FleetMaintenanceScheduleController.php
@@ -22,6 +22,7 @@ class FleetMaintenanceScheduleController extends Controller
     // View all maintenance schedules
     public function index()
     {
+        $this->authorize('viewAny', FleetMaintenanceSchedule::class);
         $schedules = FleetMaintenanceSchedule::with('vehicle','maintenanceStatus')
             ->orderByDesc('ScheduleID','desc')
             ->get();
@@ -32,6 +33,7 @@ class FleetMaintenanceScheduleController extends Controller
     // Show form to create a new schedule
     public function create()
     {
+        $this->authorize('create', FleetMaintenanceSchedule::class);
         $vehicles = FleetVehicle::all();
         $maintenanceType = CodeDetail::where('CodeID', 'FleetMaintenanceType')
             ->orderBy('Value')
@@ -43,6 +45,7 @@ class FleetMaintenanceScheduleController extends Controller
     // Store a new maintenance schedule
     public function store(FleetMaintenanceScheduleRequest $request)
     {
+        $this->authorize('create', FleetMaintenanceSchedule::class);
         $data = $request->validated();
         $this->scheduleService->create($data);
 
@@ -54,6 +57,7 @@ class FleetMaintenanceScheduleController extends Controller
     // Show form to edit a schedule
     public function edit($id)
     {
+        $this->authorize('edit', FleetMaintenanceSchedule::class);
         $schedule = FleetMaintenanceSchedule::findOrFail($id);
         $vehicles = FleetVehicle::all();
         $maintenanceType = CodeDetail::where('CodeID', 'FleetMaintenanceType')
@@ -66,6 +70,7 @@ class FleetMaintenanceScheduleController extends Controller
     // Update a schedule (acknowledge)
     public function update(FleetMaintenanceScheduleRequest $request, $id)
 {
+    $this->authorize('update', FleetMaintenanceSchedule::class);
     $data = $request->validated();
 
     // Only handle mileage update & completion
@@ -93,6 +98,7 @@ class FleetMaintenanceScheduleController extends Controller
     // Show a single schedule
     public function show($id)
     {
+        $this->authorize('view', FleetMaintenanceSchedule::class);
         $schedule = FleetMaintenanceSchedule::with(['vehicle', 'maintenanceType', 'alert'])
             ->findOrFail($id);
 
@@ -102,6 +108,7 @@ class FleetMaintenanceScheduleController extends Controller
     // Cancel a schedule
    public function cancel($id)
 {
+    $this->authorize('cancel', FleetMaintenanceSchedule::class);
     $schedule = FleetMaintenanceSchedule::findOrFail($id);
 
     // Just deactivate, no change to MaintenanceStatus
@@ -122,6 +129,7 @@ class FleetMaintenanceScheduleController extends Controller
     // Soft delete a schedule
     public function destroy($id)
     {
+        $this->authorize('destroy', FleetMaintenanceSchedule::class);
         $this->scheduleService->delete($id);
 
         return redirect()
